tweaks to how config lookup works
Additional changes:
- allow the `pool_name` to be configurable.
- delete the unused `config` directory

I'm not really sure if I like the way this commit introduces a ton of
config lookups (on each call really). Might be worth breaking out a
profiler and seeing how much time is actually spent on those calls. It
is starting to feel like using an Agent to hold the configuration state
might be a better path forward. Or forcing the `pool_name` to be set at
compile time that way we can use a macro or something to still generate
a module attribute (or something to that effect, not sure if exactly
that is possible).